+++ src/api/src/Audiochan.API/Controllers/Me/FollowingsController.cs
@@ -1,11 +1,10 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using Audiochan.API.Extensions;
+using Audiochan.Core.Artists.Commands;
+using Audiochan.Core.Artists.Queries;
 using Audiochan.Core.Common.Extensions;
 using Audiochan.Core.Common.Interfaces.Services;
-using Audiochan.Core.Users;
-using Audiochan.Core.Users.Commands;
-using Audiochan.Core.Users.Queries;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
