+++ src/api/src/Audiochan.Core/Artists/Commands/UpdateProfile/UpdateProfileRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace Audiochan.Core.Artists.Commands
+{
+    public record UpdateProfileRequest
+    {
+        public string? DisplayName { get; init; }
+    }
+}
