+++ src/api/src/Audiochan.Core/Artists/Queries/GetArtistAudios/GetUsersAudioQuery.cs
@@ -1,17 +1,29 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using Ardalis.Specification;
 using Audiochan.Core.Audios;
 using Audiochan.Core.Common.Interfaces.Pagination;
 using Audiochan.Core.Common.Interfaces.Persistence;
 using Audiochan.Core.Common.Models.Pagination;
+using Audiochan.Domain.Entities;
 using MediatR;
 
-namespace Audiochan.Core.Users.Queries
+namespace Audiochan.Core.Artists.Queries
 {
     public record GetUsersAudioQuery(string Username) : IHasOffsetPage, IRequest<OffsetPagedListDto<AudioDto>>
     {
         public int Offset { get; init; } = 1;
         public int Size { get; init; } = 30;
+    }
+    
+    public sealed class GetUserAudiosSpecification : Specification<Audio>
+    {
+        public GetUserAudiosSpecification(string username)
+        {
+            Query.AsNoTracking();
+            Query.Where(a => a.Artist.UserName == username);
+            Query.OrderByDescending(a => a.Id);
+        }
     }
 
     public class GetUsersAudioQueryHandler : IRequestHandler<GetUsersAudioQuery, OffsetPagedListDto<AudioDto>>
