+++ src/api/src/Audiochan.Core/Audios/Commands/CreateAudio/CreateAudioCommand.cs
@@ -10,6 +10,7 @@ using Audiochan.Core.Common.Interfaces.Persistence;
 using Audiochan.Core.Common.Interfaces.Services;
 using Audiochan.Core.Common.Models;
 using Audiochan.Domain.Entities;
+using FluentValidation;
 using MediatR;
 using Microsoft.Extensions.Options;
 
@@ -26,6 +27,47 @@ namespace Audiochan.Core.Audios.Commands
         public string Description { get; init; } = string.Empty;
         public List<string> Tags { get; init; } = new();
         public string BlobName => UploadId + Path.GetExtension(FileName);
+    }
+    
+    public class CreateAudioCommandValidator : AbstractValidator<CreateAudioCommand>
+    {
+        public CreateAudioCommandValidator(IOptions<MediaStorageSettings> options)
+        {
+            var storageSettings = options.Value;
+            RuleFor(req => req.UploadId)
+                .NotEmpty()
+                .WithMessage("UploadId is required.");
+            RuleFor(req => req.Duration)
+                .NotEmpty()
+                .WithMessage("Duration is required.");
+            RuleFor(req => req.FileSize)
+                .FileSizeValidation(storageSettings.Audio.MaximumFileSize);
+            RuleFor(req => req.FileName)
+                .FileNameValidation(storageSettings.Audio.ValidContentTypes);
+            RuleFor(req => req.Title)
+                .NotEmpty()
+                .WithMessage("Title is required.")
+                .MaximumLength(30)
+                .WithMessage("Title cannot be no more than 30 characters long.");
+            RuleFor(req => req.Description)
+                .NotNull()
+                .WithMessage("Description cannot be null.")
+                .MaximumLength(500)
+                .WithMessage("Description cannot be more than 500 characters long.");
+            RuleFor(req => req.Tags)
+                .NotNull()
+                .WithMessage("Tags cannot be null.")
+                .Must(u => u!.Count <= 10)
+                .WithMessage("Can only have up to 10 tags per audio upload.")
+                .ForEach(tagsRule =>
+                {
+                    tagsRule
+                        .NotEmpty()
+                        .WithMessage("Each tag cannot be empty.")
+                        .Length(3, 15)
+                        .WithMessage("Each tag must be between 3 and 15 characters long.");
+                });
+        }
     }
 
     public class CreateAudioCommandHandler : IRequestHandler<CreateAudioCommand, Result<long>>
