--- src/api/src/Audiochan.Core/Audios/Commands/CreateAudio/CreateAudioCommandValidator.cs
@@ -1,48 +1,0 @@
-﻿using Audiochan.Core.Common;
-using Audiochan.Core.Common.Extensions;
-using FluentValidation;
-using Microsoft.Extensions.Options;
-
-namespace Audiochan.Core.Audios.Commands
-{
-    public class CreateAudioCommandValidator : AbstractValidator<CreateAudioCommand>
-    {
-        public CreateAudioCommandValidator(IOptions<MediaStorageSettings> options)
-        {
-            var storageSettings = options.Value;
-            RuleFor(req => req.UploadId)
-                .NotEmpty()
-                .WithMessage("UploadId is required.");
-            RuleFor(req => req.Duration)
-                .NotEmpty()
-                .WithMessage("Duration is required.");
-            RuleFor(req => req.FileSize)
-                .FileSizeValidation(storageSettings.Audio.MaximumFileSize);
-            RuleFor(req => req.FileName)
-                .FileNameValidation(storageSettings.Audio.ValidContentTypes);
-            RuleFor(req => req.Title)
-                .NotEmpty()
-                .WithMessage("Title is required.")
-                .MaximumLength(30)
-                .WithMessage("Title cannot be no more than 30 characters long.");
-            RuleFor(req => req.Description)
-                .NotNull()
-                .WithMessage("Description cannot be null.")
-                .MaximumLength(500)
-                .WithMessage("Description cannot be more than 500 characters long.");
-            RuleFor(req => req.Tags)
-                .NotNull()
-                .WithMessage("Tags cannot be null.")
-                .Must(u => u!.Count <= 10)
-                .WithMessage("Can only have up to 10 tags per audio upload.")
-                .ForEach(tagsRule =>
-                {
-                    tagsRule
-                        .NotEmpty()
-                        .WithMessage("Each tag cannot be empty.")
-                        .Length(3, 15)
-                        .WithMessage("Each tag must be between 3 and 15 characters long.");
-                });
-        }
-    }
-}
