--- src/api/src/Audiochan.Core/Audios/Commands/CreateUpload/CreateUploadCommandValidator.cs
@@ -1,19 +1,0 @@
-﻿using Audiochan.Core.Common;
-using Audiochan.Core.Common.Extensions;
-using FluentValidation;
-using Microsoft.Extensions.Options;
-
-namespace Audiochan.Core.Audios.Commands
-{
-    public class CreateUploadCommandValidator : AbstractValidator<CreateUploadCommand>
-    {
-        public CreateUploadCommandValidator(IOptions<MediaStorageSettings> options)
-        {
-            var uploadOptions = options.Value.Audio;
-            RuleFor(req => req.FileSize)
-                .FileSizeValidation(uploadOptions.MaximumFileSize);
-            RuleFor(req => req.FileName)
-                .FileNameValidation(uploadOptions.ValidContentTypes);
-        }
-    }
-}
