--- src/api/src/Audiochan.Core/Audios/Commands/UpdateAudio/LoadAudioWithOwnerSpecification.cs
@@ -1,14 +1,0 @@
-﻿using Ardalis.Specification;
-using Audiochan.Domain.Entities;
-
-namespace Audiochan.Core.Audios.Commands
-{
-    public sealed class LoadAudioWithOwnerSpecification : Specification<Audio>
-    {
-        public LoadAudioWithOwnerSpecification(long audioId)
-        {
-            Query.Include(a => a.Artist);
-            Query.Where(a => a.Id == audioId);
-        }
-    }
-}
