--- src/api/src/Audiochan.Core/Audios/Commands/UpdateAudio/UpdateAudioCommandValidator.cs
@@ -1,41 +1,0 @@
-﻿using FluentValidation;
-
-namespace Audiochan.Core.Audios.Commands
-{
-    public class UpdateAudioCommandValidator : AbstractValidator<UpdateAudioCommand>
-    {
-        public UpdateAudioCommandValidator()
-        {
-            When(req => req.Title is not null, () =>
-            {
-                RuleFor(req => req.Title)
-                    .NotEmpty()
-                    .WithMessage("Title is required.")
-                    .MaximumLength(30)
-                    .WithMessage("Title cannot be no more than 30 characters long.");
-            });
-
-            When(req => req.Description is not null, () =>
-            {
-                RuleFor(req => req.Description)
-                    .MaximumLength(500)
-                    .WithMessage("Description cannot be more than 500 characters long.");
-            });
-
-            When(req => req.Tags is not null, () =>
-            {
-                RuleFor(req => req.Tags)
-                    .Must(u => u!.Count <= 10)
-                    .WithMessage("Can only have up to 10 tags per audio upload.")
-                    .ForEach(tagsRule =>
-                    {
-                        tagsRule
-                            .NotEmpty()
-                            .WithMessage("Each tag cannot be empty.")
-                            .Length(3, 15)
-                            .WithMessage("Each tag must be between 3 and 15 characters long.");
-                    });
-            });
-        }
-    }
-}
