+++ src/api/src/Audiochan.Core/Audios/Queries/GetAudio/GetAudioQuery.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Ardalis.Specification;
 using Audiochan.Core.Common;
 using Audiochan.Core.Common.Extensions;
 using Audiochan.Core.Common.Interfaces.Persistence;
+using Audiochan.Domain.Entities;
 using MediatR;
 using Microsoft.Extensions.Caching.Distributed;
 
@@ -11,6 +13,15 @@ namespace Audiochan.Core.Audios.Queries
 {
     public record GetAudioQuery(long Id) : IRequest<AudioDto?>
     {
+    }
+    
+    public sealed class GetAudioSpecification : Specification<Audio>
+    {
+        public GetAudioSpecification(long id)
+        {
+            Query.AsNoTracking();
+            Query.Where(x => x.Id == id);
+        }
     }
 
     public class GetAudioQueryHandler : IRequestHandler<GetAudioQuery, AudioDto?>
