--- src/api/src/Audiochan.Core/Auth/Commands/Login/LoginCommandValidator.cs
@@ -1,13 +1,0 @@
-﻿using FluentValidation;
-
-namespace Audiochan.Core.Auth.Commands
-{
-    public class LoginCommandValidator : AbstractValidator<LoginCommand>
-    {
-        public LoginCommandValidator()
-        {
-            RuleFor(x => x.Login).NotEmpty();
-            RuleFor(x => x.Password).NotEmpty();
-        }
-    }
-}
