+++ src/api/src/Audiochan.Core/Auth/Queries/GetCurrentUser/GetCurrentUserQuery.cs
@@ -1,7 +1,9 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using Ardalis.Specification;
 using Audiochan.Core.Common.Attributes;
 using Audiochan.Core.Common.Interfaces.Persistence;
+using Audiochan.Domain.Entities;
 using MediatR;
 
 namespace Audiochan.Core.Auth.Queries
@@ -9,6 +11,15 @@ namespace Audiochan.Core.Auth.Queries
     [Authorize]
     public record GetCurrentUserQuery(long UserId) : IRequest<CurrentUserDto?>
     {
+    }
+    
+    public sealed class GetCurrentUserSpecification : Specification<User>
+    {
+        public GetCurrentUserSpecification(long userId)
+        {
+            Query.AsNoTracking();
+            Query.Where(u => u.Id == userId);
+        }
     }
 
     public class GetCurrentUserQueryHandler : IRequestHandler<GetCurrentUserQuery, CurrentUserDto?>
