+++ src/api/src/Audiochan.Core/Users/Commands/UpdateEmail/UpdateEmailCommand.cs
@@ -5,6 +5,7 @@ using Audiochan.Core.Common.Extensions;
 using Audiochan.Core.Common.Interfaces.Persistence;
 using Audiochan.Core.Common.Interfaces.Services;
 using Audiochan.Core.Common.Models;
+using FluentValidation;
 using MediatR;
 
 namespace Audiochan.Core.Users.Commands
@@ -20,6 +21,16 @@ namespace Audiochan.Core.Users.Commands
             UserId = userId,
             NewEmail = request.NewEmail
         };
+    }
+    
+    public class UpdateEmailCommandValidator : AbstractValidator<UpdateEmailCommand>
+    {
+        public UpdateEmailCommandValidator()
+        {
+            RuleFor(req => req.NewEmail)
+                .NotEmpty().WithMessage("Email is required.")
+                .EmailAddress().WithMessage("Email is invalid.");
+        }
     }
 
     public class UpdateEmailCommandHandler : IRequestHandler<UpdateEmailCommand, Result>
