--- src/api/src/Audiochan.Core/Users/Commands/UpdateEmail/UpdateEmailCommandValidator.cs
@@ -1,14 +1,0 @@
-﻿using FluentValidation;
-
-namespace Audiochan.Core.Users.Commands
-{
-    public class UpdateEmailCommandValidator : AbstractValidator<UpdateEmailCommand>
-    {
-        public UpdateEmailCommandValidator()
-        {
-            RuleFor(req => req.NewEmail)
-                .NotEmpty().WithMessage("Email is required.")
-                .EmailAddress().WithMessage("Email is invalid.");
-        }
-    }
-}
