--- src/api/src/Audiochan.Core/Users/Commands/UpdatePassword/UpdatePasswordCommandValidator.cs
@@ -1,20 +1,0 @@
-﻿using Audiochan.Core.Common;
-using Audiochan.Core.Common.Extensions;
-using FluentValidation;
-using Microsoft.Extensions.Options;
-
-namespace Audiochan.Core.Users.Commands
-{
-    public class UpdatePasswordCommandValidator : AbstractValidator<UpdatePasswordCommand>
-    {
-        public UpdatePasswordCommandValidator(IOptions<IdentitySettings> options)
-        {
-            RuleFor(req => req.NewPassword)
-                .NotEmpty()
-                .WithMessage("New Password is required.")
-                .NotEqual(req => req.CurrentPassword)
-                .WithMessage("New password cannot be the same as the previous.")
-                .PasswordValidation(options.Value.PasswordSettings, "New Password");
-        }
-    }
-}
