--- src/api/src/Audiochan.Core/Users/Commands/UpdateProfile/UpdateProfileRequest.cs
@@ -1,9 +1,0 @@
-﻿namespace Audiochan.Core.Users.Commands
-{
-    public record UpdateProfileRequest
-    {
-        public string? DisplayName { get; init; }
-        public string? About { get; init; }
-        public string? Website { get; init; }
-    }
-}
