--- src/api/src/Audiochan.Core/Users/Commands/UpdateUsername/UpdateUsernameCommandValidator.cs
@@ -1,15 +1,0 @@
-﻿using Audiochan.Core.Common;
-using Audiochan.Core.Common.Extensions;
-using FluentValidation;
-using Microsoft.Extensions.Options;
-
-namespace Audiochan.Core.Users.Commands
-{
-    public class UpdateUsernameCommandValidator : AbstractValidator<UpdateUsernameCommand>
-    {
-        public UpdateUsernameCommandValidator(IOptions<IdentitySettings> options)
-        {
-            RuleFor(req => req.NewUsername).UsernameValidation(options.Value.UsernameSettings);
-        }
-    }
-}
