--- src/api/src/Audiochan.Core/Users/Queries/GetProfile/GetUserByUsernameSpecification.cs
@@ -1,14 +1,0 @@
-﻿using Ardalis.Specification;
-using Audiochan.Domain.Entities;
-
-namespace Audiochan.Core.Users.Queries
-{
-    public sealed class GetUserByUsernameSpecification : Specification<Artist>
-    {
-        public GetUserByUsernameSpecification(string username)
-        {
-            Query.AsNoTracking();
-            Query.Where(u => u.UserName == username);
-        }
-    }
-}
